feat: US-001 - Split cli.py into schema.py / store.py / cli.py
Pure code movement: no renames, no signature or behavior changes.

- schema.py: constants (CORE_FIELDS, RETRACT_TYPE, BUILTIN_FIELDS,
  VALID_FIELD_TYPES, TYPE_MAP), SchemaSQL, format_schema_help,
  load_schema, build_sql, and bundled-template helpers (SCHEMAS_DIR,
  DEFAULT_TEMPLATE, list_templates, get_template_path, read_template,
  read_template_from_path, print_templates). Imports no sibling module.
- store.py: notebook discovery (.lnb.env), id/writer helpers, flatten/
  upsert/delete, incremental_ingest (+ _IngestFenceTripped), ensure_db,
  rebuild_from_jsonl. Imports only from schema.
- cli.py: argparse construction, cmd_* handlers, print_table,
  EXAMPLE_QUERIES, main(). Imports from schema and store.

Dependency direction is strictly cli -> store -> schema. Test import
block updated to point at the new module locations; no test logic changed.
